requirements.txt
diffusers==0.32.2
fastapi==0.115.8
numpy
Pillow==11.1.0
segment_anything==1.0
torch==2.6.0
opencv-python
python-multipart
uvicorn
